add rpc calls and kate RPC examples
