+++ avail-go/src/sdk/types/types.go
@@ -7,6 +7,7 @@ import (
 type MultiAddress = types.MultiAddress
 type Call = types.Call
 type Hash = types.Hash
+type U32 = types.U32
 type ChainProperties struct {
 	IsEthereum    bool
 	SS58Format    types.U32
